release v6.0x7.3-rc0
ea09471 iqfiles: update ldc DistorCoeff for sc200ai and sc450ai
cea50d2 awb: correct  converged and tolerance paras in iq
37e0fd8 camgroup: fix camgroup iso bug
...
99539b1 rv1103b: ynr add rgeSgm_scale & loNrOut_alpha params in loNr

+++ rkaiq/algos_camgroup/rk_aiq_algo_camgroup_types.h
@@ -35,7 +35,7 @@
 typedef struct rk_aiq_singlecam_3a_result_s {
     uint8_t _camId;
     uint32_t _frameId;
-    int16_t hdrIso;
+    int hdrIso;
     // ae params
     struct {
         RKAiqAecExpInfo_t* exp_tbl;
